modify table eh_user_launch_pad_items

+++ ehcore-persist/src/main/java/com/everhomes/server/schema/tables/daos/EhUserLaunchPadItemsDao.java
@@ -64,14 +64,14 @@ public class EhUserLaunchPadItemsDao extends org.jooq.impl.DAOImpl<com.everhomes
 	/**
 	 * Fetch records that have <code>owner_id IN (values)</code>
 	 */
-	public java.util.List<com.everhomes.server.schema.tables.pojos.EhUserLaunchPadItems> fetchByOwnerId(java.lang.Byte... values) {
+	public java.util.List<com.everhomes.server.schema.tables.pojos.EhUserLaunchPadItems> fetchByOwnerId(java.lang.Long... values) {
 		return fetch(com.everhomes.server.schema.tables.EhUserLaunchPadItems.EH_USER_LAUNCH_PAD_ITEMS.OWNER_ID, values);
 	}
 
 	/**
 	 * Fetch records that have <code>user_id IN (values)</code>
 	 */
-	public java.util.List<com.everhomes.server.schema.tables.pojos.EhUserLaunchPadItems> fetchByUserId(java.lang.Byte... values) {
+	public java.util.List<com.everhomes.server.schema.tables.pojos.EhUserLaunchPadItems> fetchByUserId(java.lang.Long... values) {
 		return fetch(com.everhomes.server.schema.tables.EhUserLaunchPadItems.EH_USER_LAUNCH_PAD_ITEMS.USER_ID, values);
 	}
